refactored arxiv functions into separate module
